% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_data.R
\docType{data}
\name{holiday_data}
\alias{holiday_data}
\title{US holiday data}
\format{An object of class \code{data.frame} with 2211 rows and 3 columns.}
\usage{
holiday_data
}
\description{
Data that contains major holidays and events between 1900 and 2100,
which is useful for producing time series forecasts and analyses.
}
\examples{
holiday_data
data(holiday_data)
}
\keyword{dataset}
