% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_query.R
\name{holiday_query}
\alias{holiday_query}
\title{Holiday Query}
\usage{
holiday_query(holiday_name, holiday_years = NULL)
}
\arguments{
\item{holiday_name}{A required argument that specifies the holiday name}

\item{holiday_years}{An optional argument that selects the years for the holiday.
This argument takes four-digit numeric values. It can be a single year,
a range of years, or a vector of different years}
}
\value{
A dataframe of selected holiday dates
}
\description{
An easy query that selects a specific holiday and year or years
}
\examples{
holiday_query('new year',2018:2020)
holiday_query('mlk day', c(1990,2000,2100))

}
