% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidayR.R
\docType{package}
\name{holidayR}
\alias{holidayR}
\alias{holidayR-package}
\title{holidayR package}
\description{
This package provides a collection of major US holidays
between 1900 and 2100. It also offers easy ways to pivot the collection
by selecting specific holiday and years. Years can be a single value or
multiple selections.
}
\examples{
holiday_data
holiday_query(holiday_name,[holiday_years])

}
\author{
Mingbo Gong \email{mingbogo1@gmail.com}
}
